Refactor: Improve project structure, modularity, and packaging
This commit introduces a significant refactoring and reorganization of the insideLLMs project to align with best practices.

Key changes include:

1.  **Improved Project Layout:**
    *   Example scripts (`example_*.py`) moved to an `examples/` directory.
    *   Supporting data files (`factuality_config.yaml`, `factuality_questions.jsonl`) moved to a `data/` directory.
    *   Paths in example scripts updated accordingly.

2.  **Refactored `nlp_utils.py`:**
    *   The monolithic `nlp_utils.py` has been broken down into smaller, more focused modules within a new `insideLLMs/nlp/` subdirectory (e.g., `text_cleaning.py`, `tokenization.py`).
    *   An `insideLLMs/nlp/__init__.py` now provides a unified import interface for these NLP functions (e.g., `from insideLLMs.nlp import clean_text`).
    *   `examples/example_nlp.py` updated to use the new import paths.

3.  **Refined Core Library Structure:**
    *   `insideLLMs/utils.py` (which contained dataset loaders) renamed to `insideLLMs/dataset_utils.py` for clarity.
    *   Imports updated in `insideLLMs/runner.py`.

4.  **Added Basic Package Configuration:**
    *   A `pyproject.toml` file has been added, defining project metadata, core dependencies, and optional dependencies for `nlp` and `visualization` extras.
    *   `nlp_requirements.txt` renamed to `requirements-dev.txt` and updated to include all dependencies for a development environment.

5.  **Added `tests/` Directory:**
    *   A `tests/` directory created at the project root.
    *   A basic placeholder test file (`tests/test_basic_import.py`) added using `pytest`.
    *   `pytest` added to `requirements-dev.txt`.

6.  **Updated `README.md`:**
    *   Reflects the new directory structure for examples.
    *   Includes updated installation instructions (installing from local clone, optional dependencies via `pip install .[extra]`).
    *   A new "Development" section added, explaining how to set up a dev environment and run tests.

These changes aim to enhance the project's maintainability, scalability, and ease of use for both you and other developers.
