--- nlp_requirements.txt
@@ -1,7 +1,0 @@
-nltk>=3.6.0
-spacy>=3.0.0
-scikit-learn>=1.0.0
-gensim>=4.0.0
-matplotlib>=3.4.0
-pandas>=1.3.0
-seaborn>=0.11.0
