+++ requirements-dev.txt
@@ -1,0 +1,21 @@
+# Core
+openai>=0.27.0
+anthropic>=0.3.0
+transformers>=4.0.0
+huggingface-hub>=0.10.0
+
+# NLP
+nltk>=3.6.0
+spacy>=3.0.0
+scikit-learn>=1.0.0
+gensim>=4.0.0
+
+# Visualization
+matplotlib>=3.4.0
+pandas>=1.3.0
+seaborn>=0.11.0
+
+# Development tools
+pytest>=7.0.0
+
+# Add any other development tools here in the future (e.g., black, flake8)
